update health scripts
