+++ charts_live/health/key_numbers.json
@@ -1,4 +1,13 @@
 {
+    "vaccination_full_global": "test",
+    "vaccination_rate_africa": 22.6,
+    "vaccination_date": "18 September 2022",
+    "malaria_year": 2020,
+    "malaria_africa_total": 602020,
+    "malaria_world_total": 626868,
+    "malaria_rest_of_world_total": 24848,
+    "malaria_africa_proportion": 96,
+    "malaria_rest_of_world_proportion": 4,
     "count_86_target": 17,
     "count_86_5_target": 13
 }
